色々と修正 (#59)
* 0014 webrtc_CreateRandomString の長さ引数を size_t にするのを閉じる

- C 側の webrtc_CreateRandomString 引数を int から size_t に変更する
- Rust 側の random_string 引数を i32 から usize に変更する
- random_bytes 内の不要なキャストを削除する
- 長さ 0 と usize 値を検証するテストを追加する
- CHANGES.md に [CHANGE] として記録する

* 0024 peer_connection_interface.cc の _refcounted 直接キャストを RULES.md 準拠にするのを閉じる

- set_adm / set_audio_encoder_factory / set_audio_decoder_factory で _refcounted_get() 経由に修正する
- C ラッパーの audio ヘッダー include を追加する
- CHANGES.md に [FIX] として記録する

* 0026,0029 の issue を磨き上げる

- 0026: サイズ引数変更を int→size_t に統一しRust側影響も追記する
- 0029: CFBridgingReleaseのnull安全性を根拠にnullチェック削除へ方針転換しrefactorとして整理する

* 0026: WEBRTC_DEFINE_VECTOR 系マクロの _new / _resize のサイズ引数型を int から size_t にする

* ヘッダーの並びがおかしいのを修正

* 0028 std_string_size の戻り値型を int から size_t に変更する

- webrtc/src/webrtc_c/std.h: 宣言の戻り値型を int → size_t
- webrtc/src/webrtc_c/std.cc: 実装の static_cast<int> を撤去し size_t で返す
- src/cxxstd.rs: CxxStringRef::len() の ffi 戻り値型変更に追従し len.max(0) as usize を撤去
- CHANGES.md: develop に [CHANGE] エントリを追記

* 0029 objc_NSString_release の冗長な null チェックを削除する

* 0046: libyuv.cc の C スタイルキャストを static_cast に置き換える

* 0046: issue を close する

* 0055: peer_connection_interface.h のコメントアウト済み旧仮想関数を削除する

* 0055: issue を close する

* ちゃんとコミットされていなかった

+++ src/api/video_codec_common.rs
@@ -1712,7 +1712,7 @@ pub struct VideoFrameTypeVector {
 unsafe impl Send for VideoFrameTypeVector {}
 
 impl VideoFrameTypeVector {
-    pub fn new(size: i32) -> Self {
+    pub fn new(size: usize) -> Self {
         let raw = unsafe { ffi::webrtc_VideoFrameType_vector_new(size) };
         Self {
             raw: NonNull::new(raw)
